Move methods from provider to controller
